Vantage/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

streamlit>=1.30.0
streamlit-agraph>=0.0.45

langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.1.0
langchain-neo4j>=0.2.0
langchain-experimental>=0.3.0
langgraph>=0.2.0

neo4j>=5.0.0
faiss-cpu>=1.8.0
tavily-python>=0.3.0
langchain-tavily>=0.0.1

fpdf2>=2.7.0
duckduckgo-search>=5.0.0
